+++ example/lib/pages/async_edit_builder.page.dart
@@ -1,0 +1,54 @@
+import 'package:fetcher/fetcher.dart';
+import 'package:flutter/material.dart';
+
+class AsyncEditBuilderPage extends StatelessWidget {
+  const AsyncEditBuilderPage({super.key});
+
+  static const values = [0, 2, 4, 8, 10];
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisAlignment: MainAxisAlignment.center,
+        mainAxisSize: MainAxisSize.max,
+        children: [
+
+          // Text
+          const Text(
+            'Tap on a cell to commit value.\nThe last cell throws an error.'
+          ),
+          const SizedBox(height: 20),
+
+          // Content
+          AsyncEditBuilder<int>(
+            fetchTask: () async {
+              // Task that fetch data
+              await Future.delayed(const Duration(seconds: 2));
+              return values.first;
+            },
+            commitTask: (data) async {
+              // Task that commit data
+              await Future.delayed(const Duration(seconds: 1));
+
+              // Throw error (to test)
+              if (data == values.last) {
+                throw Exception('An error occured on commit : value is preserved');
+              }
+            },
+            builder: (context, selected, commit) {
+              return ToggleButtons(
+                isSelected: values.map((value) => value == selected).toList(growable: false),
+                onPressed: (index) => commit(values[index]),
+                children: values.map((value) {
+                  return Text(value.toString());
+                }).toList(growable: false),
+              );
+            },
+          ),
+        ],
+      ),
+    );
+  }
+}