+++ lib/fetcher.dart
@@ -2,9 +2,12 @@ library fetcher;
 
 export 'src/config/default_fetcher_config.dart';
 export 'src/config/fetcher_config.dart';
+
 export 'src/fetch_builder.dart';
 export 'src/event_fetch_builder.dart';
 export 'src/async_task_builder.dart';
+export 'src/async_edit_builder.dart';
+
 export 'src/widgets/async_form.dart';
 export 'src/widgets/clear_focus_background.dart';
 export 'src/utils/utils.dart';