+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.6
 * Add FetchBuilder.initBuilder param (to be used with fetchAtInit false)
 * New static AsyncTaskBuilder.runTask method that allows to run headless task safely
+* New AsyncEditBuilder
 
 ## 0.0.5
 * Default FetchBuilderErrorWidget now handles isDense parameter